projects/finalproject/requirements.txt
flask
openai
pydantic
pypdf
python-docx
numpy
python-dotenv
striprtf
beautifulsoup4
stripe
